Temporarily redirect `sys.stdout` before calling `input`
`input` always writes to `sys.stdout`.  I tried to simulate making it
write to a specified file stream by explicitly printing the prompt
myself and then calling `input` with no argument, but that does not
behave well with `readline` (which is one of the main reasons for
using `input`), which apparently does not know where the to stop
erasing input when the user presses backspace or Ctrl+U
(unix-line-discard).

Instead temporarily replace `sys.stdout`.